Detect potential overflows in cached_test_function
This fixes a long-standing issue that we would often uselessly re-execute
the test function in cases where we knew full well it couldn't possibly succeed.
We do this by returning a dummy standin object for a full ConjectureData instance.

This allows us to unify prescreen_buffer and cached_test_function because the
former only existed to handle cases that are now correctly handled by the latter.